support opening breadcrumbs chunks

+++ encoder/ddb_bits.h
@@ -2,7 +2,7 @@
 #ifndef __DDB_BITS_H__
 #define __DDB_BITS_H__
 
-static inline uint32_t read_bits(const char *src, uint64_t offs, uint32_t bits)
+static inline uint64_t read_bits(const char *src, uint64_t offs, uint32_t bits)
 {
     uint64_t *src_w = (uint64_t*)&src[offs >> 3];
     return (*src_w >> (offs & 7)) & ((1 << bits) - 1);
